Update groupId to org.fcrepo.api
