+++ src/main/java/org/fcrepo/api/vocabulary/Event.java
@@ -15,21 +15,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.vocabulary;
+package org.fcrepo.api.vocabulary;
+
+import org.apache.commons.rdf.api.IRI;
+import org.apache.commons.rdf.api.RDF;
+import org.apache.commons.rdf.simple.SimpleRDF;
 
 /**
- * Test the PROV Ontology Class
+ * RDF Terms from the Fedora Event Vocabulary
  * @author acoburn
  */
-public class PROVTest extends AbstractVocabularyTest {
+public class Event {
 
-    @Override
-    public String namespace() {
-        return "http://www.w3.org/ns/prov#";
-    }
+    private static RDF rdf = new SimpleRDF();
 
-    @Override
-    public Class vocabulary() {
-        return PROV.class;
+    /* Namespace */
+    public static String uri = "http://fedora.info/definitions/v4/event#";
+
+    /* Classes */
+    public static IRI ResourceCreation = rdf.createIRI(uri + "ResourceCreation");
+    public static IRI ResourceDeletion = rdf.createIRI(uri + "ResourceDeletion");
+    public static IRI ResourceModification = rdf.createIRI(uri + "ResourceModification");
+    public static IRI ResourceRelocation = rdf.createIRI(uri + "ResourceRelocation");
+
+    private Event() {
+        // prevent instantiation
     }
 }
