+++ src/test/java/org/fcrepo/api/vocabulary/PROVTest.java
@@ -15,30 +15,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.vocabulary;
-
-import org.apache.commons.rdf.api.IRI;
-import org.apache.commons.rdf.api.RDF;
-import org.apache.commons.rdf.simple.SimpleRDF;
+package org.fcrepo.api.vocabulary;
 
 /**
- * RDF Terms from the Fedora Event Vocabulary
+ * Test the PROV Ontology Class
  * @author acoburn
  */
-public class Event {
+public class PROVTest extends AbstractVocabularyTest {
 
-    private static RDF rdf = new SimpleRDF();
+    @Override
+    public String namespace() {
+        return "http://www.w3.org/ns/prov#";
+    }
 
-    /* Namespace */
-    public static String uri = "http://fedora.info/definitions/v4/event#";
-
-    /* Classes */
-    public static IRI ResourceCreation = rdf.createIRI(uri + "ResourceCreation");
-    public static IRI ResourceDeletion = rdf.createIRI(uri + "ResourceDeletion");
-    public static IRI ResourceModification = rdf.createIRI(uri + "ResourceModification");
-    public static IRI ResourceRelocation = rdf.createIRI(uri + "ResourceRelocation");
-
-    private Event() {
-        // prevent instantiation
+    @Override
+    public Class vocabulary() {
+        return PROV.class;
     }
 }
