+++ src/test/java/org/fcrepo/api/vocabulary/WebACTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.fcrepo.vocabulary;
+package org.fcrepo.api.vocabulary;
 
 /**
  * Test the Fedora WebAC Vocabulary Class
